+++ variant-switcher/src/icons/Exchange.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+
+const ExchangeIcon: React.FC = () => (
+    <svg width="11" height="12" viewBox="0 0 11 12" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path
+            d="M7 0.0839688L7 11.084M4 0.0839686L4 11.084M3.91603 0.361312L0.916031 4.36131M7.08397 10.8066L10.084 6.80662"
+            stroke="black"
+        />
+    </svg>
+);
+
+export default ExchangeIcon;